+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Font-lock: p-field variables in light color mode are darker
 - Font-lock: multiline comment in score implemented
 - Font-lock: <CsLicense> xml tag added
+- Syntax-table: multiline and single-line not conflicting
 - Repl: csoundAPI Installation script now finds the makefile
+- Repl: add instance specific environment variable options
 - Eldoc: functional syntax with specific rate, colon seperated, implemented
 
 ## [0.1.1]